feat: 서비스와 독립적인 기능들을 새로운 유틸 폴더에 추가
